pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-tracker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2",
    "httpx>=0.27",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
]

[project.scripts]
track = "pipeline.cli:cli"

[tool.setuptools.packages.find]
include = ["pipeline*"]

[tool.ruff]
line-length = 100
